backend/pyproject.toml
[tool.poetry]
name = "personal-book-tracker-backend"
version = "0.1.0"
description = "FastAPI + Strawberry GraphQL backend for the Personal Book Tracker app."
authors = ["Your Name <you@example.com>"]

packages = [
    { include = "backend", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "*"
strawberry-graphql = { extras = ["fastapi"], version = "*" }
sqlalchemy = "*"
asyncpg = "*"
uvicorn = { extras = ["standard"], version = "*" }
redis = "*"
aiobreaker = "*"
pydantic = "*"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

